API/desafio-api-produtos/api_produtos/api/urls.py
from django.urls import path, include
from django.contrib.auth.models import User
from rest_framework import routers, serializers, viewsets
from .views import Tbl_ProdutosViewSet, Tbl_InventariosViewSet


# Routers provide an easy way of automatically determining the URL conf.
router = routers.DefaultRouter()
router.register(r'produtos_api', Tbl_ProdutosViewSet)
router.register(r'inventarios_api', Tbl_InventariosViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
]