API/desafio-api-produtos/api_produtos/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
mysqlclient==2.0.3
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
